feat: #266 사용자 인증 통계 조회 API K6 스크립트
