--- k6/test/api-load-test/read/11_get-my-profile-v1.js
@@ -1,49 +1,0 @@
-import { sleep } from 'k6';
-import {check} from 'k6';
-import { SharedArray } from 'k6/data';
-import http from 'k6/http';
-import {parseResponseBody, setRequestHeader} from "../../../common/util/api-util.js";
-import {API_BASE_URL} from "../../../common/secret/secret.js";
-
-const tokens = new SharedArray('tokens', () => JSON.parse(open('../../../common/secret/tokens.json')));
-
-export function setup() {
-    console.log("⏰ 5초 대기 시작.");
-    sleep(5);
-    console.log("✅ 5초 대기 완료.");
-}
-
-export const options = {
-    setupTimeout: '30m',
-    scenarios: {
-        default: {
-            executor: 'per-vu-iterations',
-            vus: 1,
-            // vus: 100,
-            // vus: 400,
-            iterations: 1,
-            maxDuration: '30m',
-        },
-    },
-};
-
-export default function () {
-    const vuIndex = __VU - 1;
-    const response = requestApi(vuIndex);
-    const responseBody = parseResponseBody(response);
-    const responseData = responseBody.data;
-
-    check(null, {
-        'API HTTP 상태 코드 200': () => response?.status === 200,
-        'API 응답 코드 success': () => responseBody?.code === 'success',
-        '응답 데이터 - name 존재': () => responseData?.name !== undefined,
-        '응답 데이터 - profileImageUrl 존재': () => responseData?.profileImageUrl !== undefined,
-    });
-}
-
-function requestApi(vuIndex) {
-    const timeout = '1800s';
-    const headers = setRequestHeader(tokens[vuIndex]);
-
-    return http.get(`${API_BASE_URL}/api/v1/my/profile`, { headers, timeout });
-}
